Rewrote variance.py's analysis of variations
It now has a Space type to take care of transforming to subspaces and
describing the results in terms of column names.  The contrast()
report now reports all constant columns, all linear dependencies
present in all the data and any linear dependencies specific to
passing or failing cases.  It returns the sub-Space objects that
describe the ranges of actual variation of the passing and failing
data-sets.